% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage_Tian.R
\name{expit}
\alias{expit}
\title{Expit}
\usage{
expit(x)
}
\arguments{
\item{x}{data to transform}
}
\value{
vector of transformed data
}
\description{
Standard expit function
}
\details{
y = exp(x) / (1 + exp(x))
}
